namespaces change, infrastructure and async func. to internal =============================================================
System Information:
Engine Version: Atom Ver 0.22.4.20200310
Library Version:  Ver 0.19.0.20200301

OS/HW Interface:
Windows Ver 0.11.3.20200307
Linux Ver 0.11.1.20200306

Hosts:
RX Interactive Console Ver 0.16.0.20200310
RX Pipe Child Ver 0.5.0.20200310
RX Simple Host Ver 0.2.0.20200310

Storages:
File System Storage Ver 0.3.5.20200310

Compilers:
MSVC 14.24.28316
GCC 8.3.0

OSes:
Linux Ver:4.15.0-88-generic [x86_64]
Windows Workstation 10.0.18362 [x64]

+++ lib/security/rx_security.cpp
@@ -163,7 +163,7 @@ rx_security_handle_t security_manager::context_activated (security_context::smar
 	SECURITY_LOG_INFO("manager", 900, buff);
 	locks::auto_lock dummy(&active_lock_);
 	
-	uint64_t new_id;
+	intptr_t new_id;
 	do
 	{
 		if (RX_OK == rx_crypt_gen_random(&new_id, sizeof(new_id)))
